+++ src/observer/TeamComponentObserver.java
@@ -1,17 +1,18 @@
 package observer;
 
-import java.sql.Date;
+
+import java.util.Date;
 
 import model.Player;
 import model.Player.PLAYEROLE;
-import model.Staff.ROLE;
 import model.Staff;
+import model.Staff.ROLE;
 
 public interface TeamComponentObserver {
 
 	void removePlayer(Player p);
 	void removeStaff(Staff s);
-	void addPlayer(String name, String surname, PLAYEROLE role, float height, String cf, Date birth);
+	void addPlayer(String name, String surname, PLAYEROLE role, double height, String cf, Date birth);
 	void addStaff(String name, String surname, ROLE role, String cf, Date birth);
 
 }
